Simple maven skeleton
This skeleton defines 4 maven projects. One being the parent and the
other three being the children. The children consist of the client and
the server, which both depended on a single shared project.

Currently there's no other logic in this code, it's simply main classes
that demo the fucntion of the maven project structure.

Also adds a gitignore which ignores project files and maven target folders.