Demo implementation of frontend
This commit introduces a first implementation for the frontend.
It contains four major features:
  * Account balance
  * Drink consumption
  * Consumption / Balance history
  * A drink manager
For most of these, the data is currently hardcoded and can't be
changed through user interaction. The only exception is the
account balance, which can be increased / decrased by adding
balance to the account or consuming a drink.

All other features that currently don't use live-data will work
as soon as the backend contains the necessary functions.

Additionally this commit adds MigLayout 5.X as a dependency.

The children POM files now define UTF-8 as their soruce encoding