Adminview and Broadcasting Server
Drinks quantity added and the admin can set the quantity in the
adminview if he has ordered new drinks. The client now finds the server
automatically if they are in the same network. Also changed history and
added graphs to see the consumption per day. The admin can also set the
total balance and see the complete history.