No more getInstance and better UIManager
Almost all getInstance of the classes were removed and replaced by
instances which are located in the mainframe.
Undoed actions are no longer considered by the charts and estimated
remaining days. The drinkInfo button is no longer displayed when the
"really want to drink?"-Dialog is open. And the UI-Manager has
been improved, because it took some time to load the darkmode in the
login screen and the layout was broken.